uradjen prikaz studGodine za jednog studenta, izmenjane neke stvari jer je pokusano dodavanje SG prilikom dodavanja studenta i popravljena funkcija za personalni broj

+++ eObrazovanjee/src/main/java/com/ftn/eObrazovanjee/controller/StudijskaGodinaController.java
@@ -151,7 +151,7 @@ public class StudijskaGodinaController {
 	}
 	
 	@GetMapping("/tokStudija")
-	public ResponseEntity<?> studijskaGodinaStudenta(@RequestParam("idStudenta") Long idStudenta){
+	public ResponseEntity<?> studijskaGodinaStudenta(@RequestParam("idStudenta") int idStudenta){
 		try {
 			List<StudijskaGodinaDTO> response = studijskaGodinaService.StudijskaGodinaStudentaNative(idStudenta);
